Move TimeTraveledStops tests to separate file
And added tests using the AdHoc system.
